Coverity #1085: Double free of pointer "*listenObjsRet"
If malloc failed in the loop in IceListenForConnections, the error path
would free all previous allocations, then loop around and try again, and
if it failed again, free the previous allocations again.   On the other
hand, if it succeeded on the later tries, then the memory would just be
leaked, since the error would be returned and not the pointer to them.