_IceRead: Avoid possible use-after-free
`_IceRead()` gets called from multiple places which do not expect the
connection to be freed.

Do not free the connection data in `_IceRead()` to avoid potential
use-after-free issue in the various callers.

The connection data will be freed eventually in `ProcessWantToClose()`,
so not freeing it in `_IceRead()` should not introduce an memory leak.

Signed-off-by: Olivier Fourdan <ofourdan@redhat.com>